add config flag for YAML support, separate json and yaml serializers/deserializers.
