+++ plugins/dynamic-security/clients.c
@@ -27,8 +27,6 @@ Contributors:
 #include "json_help.h"
 
 #include "dynamic_security.h"
-#include "yaml_help.h"
-#include "yaml.h"
 
 struct connection_array_context{
 	const char *username;
@@ -56,7 +54,7 @@ static void client__remove_all_roles(struct dynsec__client *client);
  * #
  * ################################################################ */
 
-static int client_cmp(void *a, void *b)
+int client_cmp(void *a, void *b)
 {
 	struct dynsec__client *client_a = a;
 	struct dynsec__client *client_b = b;
@@ -128,464 +126,9 @@ void dynsec_clients__cleanup(struct dynsec__data *data)
 
 /* ################################################################
  * #
- * # Config file load and save
+ * # Command processing
  * #
  * ################################################################ */
-
-int dynsec_clients__config_load_json(struct dynsec__data *data, cJSON *tree)
-{
-	cJSON *j_clients, *j_client, *jtmp, *j_roles, *j_role;
-	cJSON *j_salt, *j_password, *j_iterations;
-	struct dynsec__client *client;
-	struct dynsec__role *role;
-	unsigned char *buf;
-	unsigned int buf_len;
-	int priority;
-	int iterations;
-	size_t username_len;
-
-	j_clients = cJSON_GetObjectItem(tree, "clients");
-	if(j_clients == NULL){
-		return 0;
-	}
-
-	if(cJSON_IsArray(j_clients) == false){
-		return 1;
-	}
-
-	cJSON_ArrayForEach(j_client, j_clients){
-		if(cJSON_IsObject(j_client) == true){
-			/* Username */
-			jtmp = cJSON_GetObjectItem(j_client, "username");
-			if(jtmp == NULL || !cJSON_IsString(jtmp)){
-				continue;
-			}
-			username_len = strlen(jtmp->valuestring);
-			if(username_len == 0){
-				continue;
-			}
-
-			client = mosquitto_calloc(1, sizeof(struct dynsec__client) + username_len + 1);
-			if(client == NULL){
-				return MOSQ_ERR_NOMEM;
-			}
-			strncpy(client->username, jtmp->valuestring, username_len);
-
-			jtmp = cJSON_GetObjectItem(j_client, "disabled");
-			if(jtmp && cJSON_IsBool(jtmp)){
-				client->disabled = cJSON_IsTrue(jtmp);
-			}
-
-			/* Salt */
-			j_salt = cJSON_GetObjectItem(j_client, "salt");
-			j_password = cJSON_GetObjectItem(j_client, "password");
-			j_iterations = cJSON_GetObjectItem(j_client, "iterations");
-
-			if(j_salt && cJSON_IsString(j_salt)
-					&& j_password && cJSON_IsString(j_password)
-					&& j_iterations && cJSON_IsNumber(j_iterations)){
-
-				iterations = (int)j_iterations->valuedouble;
-				if(iterations < 1){
-					mosquitto_free(client);
-					continue;
-				}else{
-					client->pw.iterations = iterations;
-				}
-
-				if(base64__decode(j_salt->valuestring, &buf, &buf_len) != MOSQ_ERR_SUCCESS
-						|| buf_len > sizeof(client->pw.salt)){
-
-                    mosquitto_free(buf);
-					mosquitto_free(client->username);
-					mosquitto_free(client);
-					continue;
-				}
-				memcpy(client->pw.salt, buf, (size_t)buf_len);
-				client->pw.salt_len = (size_t)buf_len;
-				mosquitto_free(buf);
-
-				if(base64__decode(j_password->valuestring, &buf, &buf_len) != MOSQ_ERR_SUCCESS
-						|| buf_len != sizeof(client->pw.password_hash)){
-
-                    mosquitto_free(buf);
-					mosquitto_free(client->username);
-					mosquitto_free(client);
-					continue;
-				}
-				memcpy(client->pw.password_hash, buf, (size_t)buf_len);
-				mosquitto_free(buf);
-				client->pw.valid = true;
-			}else{
-				client->pw.valid = false;
-			}
-
-			/* Client id */
-			jtmp = cJSON_GetObjectItem(j_client, "clientid");
-			if(jtmp != NULL && cJSON_IsString(jtmp)){
-				client->clientid = mosquitto_strdup(jtmp->valuestring);
-				if(client->clientid == NULL){
-					mosquitto_free(client);
-					continue;
-				}
-			}
-
-			/* Text name */
-			jtmp = cJSON_GetObjectItem(j_client, "textname");
-			if(jtmp != NULL && cJSON_IsString(jtmp)){
-				client->text_name = mosquitto_strdup(jtmp->valuestring);
-				if(client->text_name == NULL){
-					mosquitto_free(client->clientid);
-					mosquitto_free(client);
-					continue;
-				}
-			}
-
-			/* Text description */
-			jtmp = cJSON_GetObjectItem(j_client, "textdescription");
-			if(jtmp != NULL && cJSON_IsString(jtmp)){
-				client->text_description = mosquitto_strdup(jtmp->valuestring);
-				if(client->text_description == NULL){
-					mosquitto_free(client->text_name);
-					mosquitto_free(client->clientid);
-					mosquitto_free(client);
-					continue;
-				}
-			}
-
-			/* Roles */
-			j_roles = cJSON_GetObjectItem(j_client, "roles");
-			if(j_roles && cJSON_IsArray(j_roles)){
-				cJSON_ArrayForEach(j_role, j_roles){
-					if(cJSON_IsObject(j_role)){
-						jtmp = cJSON_GetObjectItem(j_role, "rolename");
-						if(jtmp && cJSON_IsString(jtmp)){
-							json_get_int(j_role, "priority", &priority, true, -1);
-							role = dynsec_roles__find(data, jtmp->valuestring);
-							dynsec_rolelist__client_add(client, role, priority);
-						}
-					}
-				}
-			}
-
-			HASH_ADD(hh, data->clients, username, username_len, client);
-		}
-	}
-	HASH_SORT(data->clients, client_cmp);
-
-	return 0;
-}
-
-int dynsec_clients__config_load_yaml(yaml_parser_t *parser, yaml_event_t *event, struct dynsec__data *data)
-{
-    unsigned char *buf = NULL;
-    unsigned int buf_len;
-
-    struct dynsec__client *client = NULL;
-    bool disabled = 1;
-    char* clientid;
-    char* textname;
-    char* textdescription;
-    char *salt;
-    char *pw;
-    long int iterations;
-    struct dynsec__rolelist *rolelist;
-
-    printf("%s:%d\n", __FILE__, __LINE__);
-    YAML_PARSER_SEQUENCE_FOR_ALL(parser, event, { goto error; }, {
-        printf("%s:%d\n", __FILE__, __LINE__);
-
-        client = NULL;
-        disabled = 0;
-        clientid = textname = textdescription = salt = pw = NULL;
-        iterations = 0;
-        rolelist = NULL;
-
-        printf("%s:%d\n", __FILE__, __LINE__);
-        YAML_PARSER_MAPPING_FOR_ALL(parser, event, key, { goto error; }, {
-                printf("%s:%d\n", __FILE__, __LINE__);
-            if (strcmp(key, "username") == 0) {
-                printf("%s:%d\n", __FILE__, __LINE__);
-                char *username;
-                YAML_EVENT_INTO_SCALAR_STRING(event, &username, { goto error; });
-                client = dynsec_clients__find_or_create(data, username);
-                mosquitto_free(username);
-            } else if (strcmp(key, "disabled") == 0) {
-                printf("%s:%d\n", __FILE__, __LINE__);
-                YAML_EVENT_INTO_SCALAR_BOOL(event, &disabled, { goto error; });
-            } else if (strcmp(key, "clientid") == 0) {
-                printf("%s:%d\n", __FILE__, __LINE__);
-                YAML_EVENT_INTO_SCALAR_STRING(event, &clientid, { goto error; });
-            } else if (strcmp(key, "textname") == 0) {
-                printf("%s:%d\n", __FILE__, __LINE__);
-                YAML_EVENT_INTO_SCALAR_STRING(event, &textname, { goto error; });
-            } else if (strcmp(key, "textdescription") == 0) {
-                printf("%s:%d\n", __FILE__, __LINE__);
-                YAML_EVENT_INTO_SCALAR_STRING(event, &textdescription, { goto error; });
-            } else if (strcmp(key, "salt") == 0) {
-                printf("%s:%d\n", __FILE__, __LINE__);
-                YAML_EVENT_INTO_SCALAR_STRING(event, &salt, { goto error; });
-            } else if (strcmp(key, "password") == 0) {
-                printf("%s:%d\n", __FILE__, __LINE__);
-                YAML_EVENT_INTO_SCALAR_STRING(event, &pw, { goto error; });
-            } else if (strcmp(key, "iterations") == 0) {
-                printf("%s:%d\n", __FILE__, __LINE__);
-                YAML_EVENT_INTO_SCALAR_LONG_INT(event, &iterations, { goto error; });
-            } else if (strcmp(key, "roles") == 0) {
-                if (dynsec_rolelist__load_from_yaml(parser, event, data, &rolelist)) goto error;
-
-                printf("%s:%d\n", __FILE__, __LINE__);
-            } else {
-                yaml_dump_block(parser, event);
-                printf("%s:%d\n", __FILE__, __LINE__);
-                mosquitto_log_printf(MOSQ_LOG_ERR, "Unexpected key for client config %s \n", key);
-            }
-                printf("%s:%d\n", __FILE__, __LINE__);
-        });
-
-        if (client) {
-            client->clientid = clientid;
-            client->text_name = textname;
-            client->text_description = textdescription;
-            client->disabled = disabled;
-
-            if (rolelist) {
-                struct dynsec__rolelist *iter;
-                struct dynsec__rolelist *tmp;
-
-                printf("%s:%d\n", __FILE__, __LINE__);
-                HASH_ITER(hh, rolelist, iter, tmp){
-                    printf("%s:%d\n", __FILE__, __LINE__);
-                    dynsec_rolelist__add(&client->rolelist, iter->role, iter->priority);
-                    dynsec_clientlist__add(&iter->role->clientlist, client, iter->priority);
-                    iter->role = NULL;
-                    printf("%s:%d\n", __FILE__, __LINE__);
-                }
-
-                dynsec_rolelist__cleanup(&rolelist);
-            }
-
-            if (salt && pw && iterations > 0) {
-                printf("PW VALID FOR %s\n", client->username);
-                client->pw.valid = 1;
-                client->pw.iterations = (int)iterations;
-
-                if(base64__decode(salt, &buf, &buf_len) == MOSQ_ERR_SUCCESS && buf_len == sizeof(client->pw.salt)) {
-                    memcpy(client->pw.salt, buf, (size_t)buf_len);
-                    mosquitto_free(buf);
-                    buf = NULL;
-                } else {
-                    client->pw.valid = 0;
-                    mosquitto_free(buf);
-                    buf = NULL;
-                }
-
-                if(base64__decode(pw, &buf, &buf_len) == MOSQ_ERR_SUCCESS && buf_len == sizeof(client->pw.password_hash)) {
-                    memcpy(client->pw.password_hash, buf, (size_t)buf_len);
-                    mosquitto_free(buf);
-                    buf = NULL;
-                } else {
-                    mosquitto_free(buf);
-                    buf = NULL;
-                    client->pw.valid = 0;
-                }
-            } else {
-                printf("PW NOT VALID FOR %s\n", client->username);
-                client->pw.valid = 0;
-            }
-
-            mosquitto_free(salt);
-            mosquitto_free(pw);
-
-        } else {
-            mosquitto_free(clientid);
-            mosquitto_free(textname);
-            mosquitto_free(textdescription);
-            dynsec_rolelist__cleanup(&rolelist);
-        }
-
-printf("%s:%d\n", __FILE__, __LINE__);
-
-    });
-
-    printf("%s:%d\n", __FILE__, __LINE__);
-
-    return 1;
-error:
-    mosquitto_free(clientid);
-    mosquitto_free(textname);
-    mosquitto_free(textdescription);
-    dynsec_rolelist__cleanup(&rolelist);
-    mosquitto_free(salt);
-    mosquitto_free(pw);
-    return 0;
-}
-
-static int dynsec__config_add_clients_yaml(yaml_emitter_t* emitter, yaml_event_t* event, struct dynsec__data *data)
-{
-
-
-    struct dynsec__client *client, *client_tmp;
-    char *buf;
-    printf("%s:%d\n", __FILE__, __LINE__);
-
-    yaml_sequence_start_event_initialize(event, NULL, (yaml_char_t *)YAML_SEQ_TAG,
-                                         1, YAML_ANY_SEQUENCE_STYLE);
-    if (!yaml_emitter_emit(emitter, event)) return 1;
-    printf("%s:%d\n", __FILE__, __LINE__);
-
-    HASH_ITER(hh, data->clients, client, client_tmp){
-        printf("%s:%d\n", __FILE__, __LINE__);
-        yaml_mapping_start_event_initialize(event, NULL, (yaml_char_t *)YAML_MAP_TAG,
-                                            1, YAML_ANY_MAPPING_STYLE);
-        if (!yaml_emitter_emit(emitter, event)) return 1;
-
-        if (!yaml_emit_string_field(emitter, event, "username", client->username)) return 1;
-
-        printf("%s:%d\n", __FILE__, __LINE__);
-        if (client->clientid && !yaml_emit_string_field(emitter, event, "clientid", client->clientid)) return 1;
-        printf("%s:%d\n", __FILE__, __LINE__);
-        if (client->text_name && !yaml_emit_string_field(emitter, event, "textname", client->text_name)) return 1;
-        printf("%s:%d\n", __FILE__, __LINE__);
-        if (client->text_description && !yaml_emit_string_field(emitter, event, "textdescription", client->text_description)) return 1;
-        printf("%s:%d\n", __FILE__, __LINE__);
-        if (client->disabled && !yaml_emit_string_field(emitter, event, "disabled", "true")) return 1;
-        printf("%s:%d\n", __FILE__, __LINE__);
-
-
-        yaml_scalar_event_initialize(event, NULL, (yaml_char_t *)YAML_STR_TAG,
-                                     (yaml_char_t *)"roles", strlen("roles"), 1, 0, YAML_PLAIN_SCALAR_STYLE);
-        if (!yaml_emitter_emit(emitter, event)) return 1;
-
-        printf("%s:%d\n", __FILE__, __LINE__);
-        if (dynsec_rolelist__all_to_yaml(client->rolelist, emitter, event)) return 1;
-        printf("%s:%d\n", __FILE__, __LINE__);
-
-        if(client->pw.valid){
-            if(base64__encode(client->pw.password_hash, sizeof(client->pw.password_hash), &buf) != MOSQ_ERR_SUCCESS){
-                mosquitto_log_printf(MOSQ_LOG_ERR, "dynsec: error encoding password hash to base64");
-                return 1;
-            }
-
-            if (!yaml_emit_string_field(emitter, event, "password", buf)) {
-                mosquitto_free(buf);
-                return 1;
-            }
-
-            mosquitto_free(buf);
-
-            if(base64__encode(client->pw.salt, sizeof(client->pw.salt), &buf) != MOSQ_ERR_SUCCESS){
-                mosquitto_log_printf(MOSQ_LOG_ERR, "dynsec: error encoding password salt to base64");
-                return 1;
-            }
-
-            if (!yaml_emit_string_field(emitter, event, "salt", buf)) {
-                mosquitto_free(buf);
-                return 1;
-            }
-
-            mosquitto_free(buf);
-
-            if (!yaml_emit_int_field(emitter, event, "iterations", client->pw.iterations)) {
-                return 1;
-            }
-        }
-
-        yaml_mapping_end_event_initialize(event);
-        if (!yaml_emitter_emit(emitter, event)) return 1;
-
-    }
-
-    printf("%s:%d\n", __FILE__, __LINE__);
-
-    yaml_sequence_end_event_initialize(event);
-    if (!yaml_emitter_emit(emitter, event)) return 1;
-
-
-    return 0;
-}
-
-
-static int dynsec__config_add_clients_json(struct dynsec__data *data, cJSON *j_clients)
-{
-	struct dynsec__client *client, *client_tmp;
-	cJSON *j_client, *j_roles, *jtmp;
-	char *buf;
-
-	HASH_ITER(hh, data->clients, client, client_tmp){
-		j_client = cJSON_CreateObject();
-		if(j_client == NULL) return 1;
-		cJSON_AddItemToArray(j_clients, j_client);
-
-		if(cJSON_AddStringToObject(j_client, "username", client->username) == NULL
-				|| (client->clientid && cJSON_AddStringToObject(j_client, "clientid", client->clientid) == NULL)
-				|| (client->text_name && cJSON_AddStringToObject(j_client, "textname", client->text_name) == NULL)
-				|| (client->text_description && cJSON_AddStringToObject(j_client, "textdescription", client->text_description) == NULL)
-				|| (client->disabled && cJSON_AddBoolToObject(j_client, "disabled", true) == NULL)
-				){
-
-			return 1;
-		}
-
-		j_roles = dynsec_rolelist__all_to_json(client->rolelist);
-		if(j_roles == NULL){
-			return 1;
-		}
-		cJSON_AddItemToObject(j_client, "roles", j_roles);
-
-		if(client->pw.valid){
-			if(base64__encode(client->pw.password_hash, sizeof(client->pw.password_hash), &buf) != MOSQ_ERR_SUCCESS){
-				return 1;
-			}
-			jtmp = cJSON_CreateString(buf);
-			mosquitto_free(buf);
-			if(jtmp == NULL) return 1;
-			cJSON_AddItemToObject(j_client, "password", jtmp);
-
-			if(base64__encode(client->pw.salt, client->pw.salt_len, &buf) != MOSQ_ERR_SUCCESS){
-				return 1;
-			}
-
-			jtmp = cJSON_CreateString(buf);
-			mosquitto_free(buf);
-			if(jtmp == NULL) return 1;
-			cJSON_AddItemToObject(j_client, "salt", jtmp);
-
-			if(cJSON_AddIntToObject(j_client, "iterations", client->pw.iterations) == NULL){
-				return 1;
-			}
-		}
-	}
-
-	return 0;
-}
-
-int dynsec_clients__config_save_yaml(yaml_emitter_t* emitter, yaml_event_t* event, struct dynsec__data *data)
-{
-    yaml_scalar_event_initialize(event, NULL, (yaml_char_t *)YAML_STR_TAG,
-                                 (yaml_char_t *)"clients", strlen("clients"), 1, 0, YAML_PLAIN_SCALAR_STYLE);
-    if (!yaml_emitter_emit(emitter, event)) return 1;
-
-    if(dynsec__config_add_clients_yaml(emitter, event, data)) return 1;
-
-    return 0;
-}
-
-
-int dynsec_clients__config_save_json(struct dynsec__data *data, cJSON *tree)
-{
-	cJSON *j_clients;
-
-	if((j_clients = cJSON_AddArrayToObject(tree, "clients")) == NULL){
-		return 1;
-	}
-	if(dynsec__config_add_clients_json(data, j_clients)){
-		return 1;
-	}
-
-	return 0;
-}
-
 
 int dynsec_clients__process_create(struct dynsec__data *data, struct control_cmd *cmd, struct mosquitto *context)
 {
