+++ plugins/dynamic-security/rolelist.c
@@ -29,7 +29,6 @@ Contributors:
 #include "json_help.h"
 #include "mosquitto.h"
 #include "mosquitto_broker.h"
-#include "yaml_help.h"
 
 
 /* ################################################################
@@ -170,143 +169,3 @@ int dynsec_rolelist__group_add(struct dynsec__group *group, struct dynsec__role 
 	return rc;
 }
 
-
-int dynsec_rolelist__load_from_json(struct dynsec__data *data, cJSON *command, struct dynsec__rolelist **rolelist)
-{
-	cJSON *j_roles, *j_role, *j_rolename;
-	int priority;
-	struct dynsec__role *role;
-
-	j_roles = cJSON_GetObjectItem(command, "roles");
-	if(j_roles){
-		if(cJSON_IsArray(j_roles)){
-			cJSON_ArrayForEach(j_role, j_roles){
-				j_rolename = cJSON_GetObjectItem(j_role, "rolename");
-				if(j_rolename && cJSON_IsString(j_rolename)){
-					json_get_int(j_role, "priority", &priority, true, -1);
-					role = dynsec_roles__find(data, j_rolename->valuestring);
-					if(role){
-						dynsec_rolelist__add(rolelist, role, priority);
-					}else{
-						dynsec_rolelist__cleanup(rolelist);
-						return MOSQ_ERR_NOT_FOUND;
-					}
-				}else{
-					return MOSQ_ERR_INVAL;
-				}
-			}
-			return MOSQ_ERR_SUCCESS;
-		}else{
-			return MOSQ_ERR_INVAL;
-		}
-	}else{
-		return ERR_LIST_NOT_FOUND;
-	}
-}
-
-int dynsec_rolelist__load_from_yaml(yaml_parser_t *parser, yaml_event_t *event, struct dynsec__data *data, struct dynsec__rolelist **rolelist)
-{
-
-    YAML_PARSER_SEQUENCE_FOR_ALL(parser, event, { goto error; }, {
-        printf("%s:%d\n", __FILE__, __LINE__);
-        char* rolename = NULL;
-        long int priority = -1;
-
-        printf("%s:%d\n", __FILE__, __LINE__);
-        YAML_PARSER_MAPPING_FOR_ALL(parser, event, key, { goto error; }, {
-                printf("%s:%d\n", __FILE__, __LINE__);
-                if (strcmp(key, "rolename") == 0) {
-                    printf("%s:%d\n", __FILE__, __LINE__);
-                    YAML_EVENT_INTO_SCALAR_STRING(event, &rolename, { goto error; });
-                } else if (strcmp(key, "priority") == 0) {
-                    printf("%s:%d\n", __FILE__, __LINE__);
-                    YAML_EVENT_INTO_SCALAR_LONG_INT(event, &priority, { goto error; });
-                } else {
-                    printf("%s:%d\n", __FILE__, __LINE__);
-                    mosquitto_log_printf(MOSQ_LOG_ERR, "Unexpected key for role config %s \n", key);
-                    yaml_dump_block(parser, event);
-                }
-        });
-
-        printf("%s:%d\n", __FILE__, __LINE__);
-
-        if (rolename) {
-            printf("rn = %s\n", rolename);
-            struct dynsec__role *role = dynsec_roles__find_or_create(data, rolename);
-            if (role) {
-                printf("%s:%d\n", __FILE__, __LINE__);
-                dynsec_rolelist__add(rolelist, role, (int)priority);
-            } else {
-                printf("OUT OF MEMORY %s:%d\n", __FILE__, __LINE__);
-                free(rolename);
-                goto error;
-            }
-        }
-
-        printf("%s:%d\n", __FILE__, __LINE__);
-    });
-
-    printf("%s:%d\n", __FILE__, __LINE__);
-
-    return 0;
-error:
-    dynsec_rolelist__cleanup(rolelist);
-    return 1;
-}
-
-
-int dynsec_rolelist__all_to_yaml(struct dynsec__rolelist *base_rolelist, yaml_emitter_t *emitter, yaml_event_t *event)
-{
-    struct dynsec__rolelist *rolelist, *rolelist_tmp;
-
-    yaml_sequence_start_event_initialize(event, NULL, (yaml_char_t *)YAML_SEQ_TAG,
-                                         1, YAML_ANY_SEQUENCE_STYLE);
-    if (!yaml_emitter_emit(emitter, event)) return 1;
-
-
-    HASH_ITER(hh, base_rolelist, rolelist, rolelist_tmp){
-
-        yaml_mapping_start_event_initialize(event, NULL, (yaml_char_t *)YAML_MAP_TAG,
-                                            1, YAML_ANY_MAPPING_STYLE);
-        if (!yaml_emitter_emit(emitter, event)) return 1;
-
-        if (!yaml_emit_string_field(emitter, event, "rolename", rolelist->role->rolename)) return 1;
-        if (rolelist->priority != -1 && !yaml_emit_int_field(emitter, event, "priority", rolelist->priority)) return 1;
-
-        yaml_mapping_end_event_initialize(event);
-        if (!yaml_emitter_emit(emitter, event)) return 1;
-    }
-
-    yaml_sequence_end_event_initialize(event);
-    if (!yaml_emitter_emit(emitter, event)) return 1;
-
-    return 0;
-}
-
-
-cJSON *dynsec_rolelist__all_to_json(struct dynsec__rolelist *base_rolelist)
-{
-	struct dynsec__rolelist *rolelist, *rolelist_tmp;
-	cJSON *j_roles, *j_role;
-
-	j_roles = cJSON_CreateArray();
-	if(j_roles == NULL) return NULL;
-
-	HASH_ITER(hh, base_rolelist, rolelist, rolelist_tmp){
-		j_role = cJSON_CreateObject();
-		if(j_role == NULL){
-			cJSON_Delete(j_roles);
-			return NULL;
-		}
-		cJSON_AddItemToArray(j_roles, j_role);
-
-		if(cJSON_AddStringToObject(j_role, "rolename", rolelist->role->rolename) == NULL
-				|| (rolelist->priority != -1 && cJSON_AddIntToObject(j_role, "priority", rolelist->priority) == NULL)
-				){
-
-			cJSON_Delete(j_roles);
-			return NULL;
-		}
-	}
-	return j_roles;
-}
